Character Sheet
The first interactive piece of code and where you begin to build your PlayerCharacter. 
I figured out a lot of cool stuff but in the end, nothing saves, nothing really happens at all.
Eventually this should have the option to choose a race, then a class, assign points for your attributes, and determine what your character looks like(maybe a backstory option?). You should be able to "interact" with NPCs and use your skills as well as gear and weapons.