Battle Simulation
***Simulation****
This takes the players armor class[AC](chance to be hit) and the enemies armor class(chance to hit) and compares them to the diceRoll. 
If the diceRoll is greater than the AC of the character then eventually it will roll to see how much damage is done(ex. sword does d6 damage), but for now it just takes the weapon damage and subtracts it from the health of character.